% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulka_functions.R
\name{n_pct}
\alias{n_pct}
\title{Categorical Table Function Arranged by Percentage Value}
\usage{
n_pct(df, var1)
}
\arguments{
\item{df}{A dataframe}

\item{var1}{A variable}
}
\description{
This function allows you to make a kabled table of count and pct
}
\examples{
n_pct(df, var1)
}
\keyword{count,}
\keyword{percent}
\keyword{table,}
