% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulka_functions.R
\name{n_pct_fct}
\alias{n_pct_fct}
\title{Categorical Table Function Arranged by Factor Levels}
\usage{
n_pct_fct(df, var1)
}
\arguments{
\item{df}{A dataframe}

\item{var1}{A factor variable}
}
\description{
This function allows you to make a kabled table of count and pct and arrange it by the factor levels
}
\examples{
n_pct(df, var1)
}
\keyword{count,}
\keyword{percent}
\keyword{table,}
