% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulka_functions.R
\name{shelve}
\alias{shelve}
\title{Continuous Table Function}
\usage{
shelve(df, var1, var2)
}
\arguments{
\item{x}{A dataframe}

\item{y}{A grouping variable}

\item{z}{A continuous variable}
}
\description{
This function allows you to make a kabled table of Q1, mean, median, Q2, max
}
\examples{
n_pct(df, var1, var2)
}
\keyword{Q1,}
\keyword{Q2,}
\keyword{max}
\keyword{mean,}
\keyword{median,}
\keyword{table,}
